jacinta: BCD-pack numbers in Int / Long / Array[Double] AST nodes (#1137)
* jacinta: store Bcd nibbles in Array[Int] instead of Array[Long]

Bcd now packs 8 nibbles per Int word (vs 16 per Long), with the
header word's sign bit at bit 31 and the nibble count in bits 0–30
(capping at ~2.1 billion nibbles per number, still effectively
unlimited for any practical JSON input).

Saves 4 bytes per Bcd instance — the header was previously a full
Long even though it only needed 33 bits (1 sign + 32 count). All
internal pattern matches and the isBcd runtime type check are
updated from Array[Long] ([J) to Array[Int] ([I).

* jacinta: add Array[Long] as BCD-packed number-array variant

Adds Array[Long] as a new JsonArray representation alongside the
existing Array[Double] and boxed IArray[Any] forms. Each Long holds
one number in a self-contained BCD encoding:

  - bit 63    : sign
  - bits 56–62: nibble count (max 14)
  - bits 0–55 : up to 14 nibbles, oldest at the highest used position

Encoding helpers (Bcd.packBcdLong, bcdLongText, bcdLongToDouble) live
in the Bcd companion. Json.Ast.bcdArr(values: Array[Long]) constructs
the variant; isNumberArray, arrayLength, arrayElement, hashCode,
equals, and the printer all recognise the new shape.

This is a type-level addition. The parser is not yet wired to emit
Array[Long] — a follow-up will switch it to prefer this representation
for arrays whose elements all fit the 14-nibble limit, skipping the
per-number Double conversion.

* jacinta: add Int as small-BCD JsonNumber variant

Adds Int as a fourth JsonNumber alongside Long, Double, and Bcd, for
numbers that fit 7 BCD nibbles (one Int) with a self-contained
sign + count + nibbles encoding:

  - bit 31    : sign
  - bits 28–30: nibble count (max 7)
  - bits 0–27 : up to 7 nibbles, oldest at highest used position

Type-level addition; parser doesn't yet emit this variant. The boxed
runtime class is `java.lang.Integer`, distinct from `java.lang.Long`,
so the type test cleanly separates the two.

`isSmallBcd` extension method, `Bcd.bcdIntText` / `bcdIntToDouble` /
`packBcdInt` helpers, and the `double` / `long` / `bcd` accessor
extensions all handle the new variant. The printer renders via the
canonical text form.

Array[Int] as a number-array variant was considered but deferred — it
shares a runtime class with Bcd and the distinction would require
content-based dispatch.

* jacinta: move Bcd to Array[Short] and open Array[Int] for small-BCD arrays

The runtime class of `Bcd` was `Array[Int]` (`[I`), which made it
indistinguishable from the natural representation for an unboxed array
of small (≤7-nibble) BCD numbers. Repack `Bcd` as `Array[Short]` (`[S`)
— 4 nibbles per word instead of 8 — to free up the `[I` slot for the
array variant, and add `Array[Int]` as a fourth `JsonArray` shape:

  - `[D` (Array[Double])  — numbers as IEEE-754 doubles
  - `[J` (Array[Long])    — numbers as single-Long BCD (14 nibbles)
  - `[I` (Array[Int])     — numbers as single-Int small-BCD (7 nibbles)
  - `[S` (Array[Short])   — a single high-precision `Bcd` value

`Json.Ast.smallBcdArr(values: Array[Int])` constructs the new variant;
`isBcdIntArray`, `arrayLength`, `arrayElement`, `hashCode`, `equals`,
and the printer all recognise it. The header word of `Bcd` shrinks
from 32 bits to 16 (1 sign bit + 15 count bits, max 32767 nibbles per
number, still effectively unlimited).

Parser still emits `Array[Double]` for number arrays — wiring it to
prefer the new shapes is a follow-up.

* jacinta: pack Bcd nibbles into Double bits via raw-bit conversion

`Bcd` storage now uses `Array[Double]` (`[D` at runtime), with each
Double's 52 mantissa bits holding 13 BCD nibbles. The exponent is
pinned to the IEEE-754 bias (`0x3FF`), so every word is a finite
double in [1.0, 2.0) — never a NaN, so
`Double.doubleToRawLongBits ∘ Double.longBitsToDouble` round-trips
exactly on any JVM. This keeps `Bcd` runtime-distinct from
`Array[Int]` / `Array[Long]` (the small/medium BCD-packed number-array
variants) while recovering most of the throughput lost by the brief
`Array[Short]` experiment.

`parseValue` and `parseNumber` gain a `bcdOnly` parameter. When set
(by `parseArray` for each element), `parseNumber` short-circuits past
the decode loop and the Double-materialisation, returning a
single-Long BCD packing directly. Without this, the parser would
build two representations per number for array elements; with it,
only the BCD form is computed.

`Array[Double]` is dropped as a `JsonArray` variant. Number-only
arrays now go to `Array[Int]` (≤7 nibbles per element) or
`Array[Long]` (≤14 nibbles), with one-step `Int → Long → boxed`
migrations on overflow. Removes the `numItems`/`numberBuffer` pool
and the `numberAsDouble`/`numberFitsDouble` side channels —
`parseArray` classifies each element directly from `parseValue`'s
return.

Throughput on `mill jacinta.bench.run`:

  Example 7 (1000 small integers):  10.2 µs → 9.7 µs  (-5%)
  Example 8 (1000 small decimals):  18.0 µs → 15.5 µs (-14%)
  Example 6 Full (high-precision):  25 µs   → 28 µs   (+13%,
    from the Bcd storage shift — Array[Double] holds 13 nibbles per
    word vs the original Array[Long]'s 16, so more array writes per
    Bcd allocation. Bcd / Double `NumberMode`s on Ex6 are unaffected.)

* ypsiloid: update Bcd pattern matches for Array[Double] storage

Yaml's AST reaches into jacinta's `Bcd` opaque type via its runtime
class, so the recent `Bcd: Array[Short] -> Array[Double]` move broke
ypsiloid's isBcd / equality / accessor pattern matches. Switch the
ypsiloid-side matches from `Array[Long]` (the original Bcd shape) to
`Array[Double]` (the current one), and update the Ex6 bench-row label
to match.

+++ lib/ypsiloid/src/core/ypsiloid.Yaml.scala
@@ -573,7 +573,7 @@ object Yaml extends Yaml2, Dynamic:
     // value that overflowed `Long`/`Double` precision during parsing.
     object BcdValue:
       def unapply(ast: Yaml.Ast): Option[Bcd] = ast match
-        case b: Array[Long] @unchecked => Some(b.asInstanceOf[Bcd])
+        case b: Array[Double] @unchecked => Some(b.asInstanceOf[Bcd])
         case _                         => None
   
     object Str:
@@ -622,19 +622,19 @@ object Yaml extends Yaml2, Dynamic:
         // Cross-shape numeric comparisons: a `Bcd` may be equal in value
         // to a `Long`, `Double`, or another `Bcd`, when their canonical
         // BigDecimal projections compare equal.
-        case (a: Array[Long] @unchecked, b: Array[Long] @unchecked) =>
+        case (a: Array[Double] @unchecked, b: Array[Double] @unchecked) =>
           a.asInstanceOf[Bcd].toBigDecimal == b.asInstanceOf[Bcd].toBigDecimal
 
-        case (a: Array[Long] @unchecked, b: Long) =>
+        case (a: Array[Double] @unchecked, b: Long) =>
           a.asInstanceOf[Bcd].toBigDecimal == BigDecimal(b)
 
-        case (a: Array[Long] @unchecked, b: Double) =>
+        case (a: Array[Double] @unchecked, b: Double) =>
           a.asInstanceOf[Bcd].toBigDecimal == BigDecimal(b)
 
-        case (a: Long, b: Array[Long] @unchecked) =>
+        case (a: Long, b: Array[Double] @unchecked) =>
           BigDecimal(a) == b.asInstanceOf[Bcd].toBigDecimal
 
-        case (a: Double, b: Array[Long] @unchecked) =>
+        case (a: Double, b: Array[Double] @unchecked) =>
           BigDecimal(a) == b.asInstanceOf[Bcd].toBigDecimal
   
         case (a: Array[AnyRef] @unchecked, b: Array[AnyRef] @unchecked) =>
@@ -660,7 +660,7 @@ object Yaml extends Yaml2, Dynamic:
       case d: Double    => d.hashCode
       case s: String    => s.hashCode
   
-      case b: Array[Long] @unchecked =>
+      case b: Array[Double] @unchecked =>
         // Hash via the BigDecimal projection so a `Bcd` whose value equals
         // a numeric `Long`/`Double` literal has a consistent hash.
         b.asInstanceOf[Bcd].toBigDecimal.hashCode
@@ -1007,7 +1007,7 @@ object Yaml extends Yaml2, Dynamic:
       case _: Double                  => YamlPrimitive.Decimal
       // High-precision BCD numbers report as `Decimal`. The AST-level
       // distinction (`isBcd`) remains available for callers that care.
-      case _: Array[Long] @unchecked  => YamlPrimitive.Decimal
+      case _: Array[Double] @unchecked  => YamlPrimitive.Decimal
       case _: String                  => YamlPrimitive.Str
 
       case xs: Array[AnyRef] @unchecked =>
